+++ database/migrations/2026_03_10_052541_add_fine_date_to_workforce_fines_table.php
@@ -9,9 +9,7 @@ class AddFineDateToWorkforceFinesTable extends Migration
     public function up(): void
     {
         if (Schema::hasTable('workforce_fines')) {
-
             if (!Schema::hasColumn('workforce_fines', 'fine_date')) {
-
                 Schema::table('workforce_fines', function (Blueprint $table) {
                     $table->date('fine_date')->nullable()->after('employee_id');
                 });
@@ -21,8 +19,12 @@ class AddFineDateToWorkforceFinesTable extends Migration
 
     public function down(): void
     {
-        Schema::table('workforce_fines', function (Blueprint $table) {
-            $table->dropColumn('fine_date');
-        });
+        if (Schema::hasTable('workforce_fines')) {
+            if (Schema::hasColumn('workforce_fines', 'fine_date')) {
+                Schema::table('workforce_fines', function (Blueprint $table) {
+                    $table->dropColumn('fine_date');
+                });
+            }
+        }
     }
 }
